refactor: align quickconfig presentation package with its directory
Keep Quick Config Compose screens in quickconfig.presentation.
Shared scaffolds stay in ui.compose.
